Use numpy broadcast to calculate iou, significantly accelerated

test in `time python example.py`
old version:
```
Iteration: 69
Accuracy: 84.23%
python example.py  161.46s user 2.53s system 94% cpu 2:53.73 total
```
new version
```
Iteration: 107
Accuracy: 84.28%
python example.py  67.12s user 3.49s system 283% cpu 24.925 total
```